+++ Server/BackOfficeDashboard/src/main/java/com/majorMedia/BackOfficeDashboard/exception/EntityNotFoundException.java
@@ -1,7 +1,7 @@
 package com.majorMedia.BackOfficeDashboard.exception;
 
 public class EntityNotFoundException extends  RuntimeException{
-    public EntityNotFoundException(Integer id, Class<?> entity) {
+    public EntityNotFoundException(Long id, Class<?> entity) {
         super(" The " + entity.getSimpleName().toLowerCase() + " with id '" + id + "' does not exist in our records");
     }
     public EntityNotFoundException(Class<?> entity) {
