+++ Server/BackOfficeDashboard/src/main/java/com/majorMedia/BackOfficeDashboard/service/adminService/IAdminService.java
@@ -18,8 +18,8 @@ import java.util.List;
 public interface IAdminService {
 
     public String updateAccountSettings(UpdateAccountRequest request);
-    public String uploadAdminAvatar(Integer adminId, MultipartFile file) throws IOException;
-    public byte[] getImageData(Integer adminId);
+    public String uploadAdminAvatar(Long adminId, MultipartFile file) throws IOException;
+    public byte[] getImageData(Long adminId);
     public String changePassword(ResetPasswordRequest resetPasswordRequest);
     public List<UserResponse> getAllOwners(String sortBy , String searchKey, String filterByProfile, String filterByStatus);
